+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 * Remove Amazon nudges on search and product detail pages
 * Send user requested URL in ECM Bot API request into header `ecm-bot-req-url`
 * [AMA] PDP, get price from HTML hidden input field
-* [AMA] PDP, get avg. rating from popover 
+* [AMA] PDP, get avg. rating from HTML popover 
+* [AMA] PDP, get count of reviews from HTML review text